inference_ray/pyproject.toml
[project]
name = "inference_ray"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.70.0",
    "imageio>=2.37.0",
    "interface",
    "data",
    "utils",
    "msgpack>=1.1.0",
    "packaging>=24.2",
    "pip>=25.0.1",
    "protobuf>=5.29.3",
    "qdrant-client>=1.13.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.48.3",
    "valkey>=6.1.0",
    "h5py>=3.12.1",
    "imageio[ffmpeg,pyav]>=2.37.0",
    "librosa>=0.11.0",
    "onnx>=1.17.0",
    "onnxruntime-gpu>=1.21.1",
    "opencv-python-headless>=4.11.0.86",
    "qdrant-client>=1.13.2",
    "ray[serve,serve-grpc]==2.42.0",
    "scipy>=1.15.2",
]

[tool.uv]
package = true

[tool.uv.sources]
interface = { workspace = true }
data = { workspace = true }
utils = { workspace = true }
